+++ qutip/solver/integrator/integrator.py
@@ -238,85 +238,3 @@ class Integrator:
                if key in new_options and new_options[key] is not None
             }
         }
-
-
-class SIntegrator(Integrator):
-    """
-    A wrapper around stochastic ODE solvers.
-
-    Parameters
-    ----------
-    system: qutip.StochasticSystem
-        Quantum system in which states evolve.
-
-    options: dict
-        Options for the integrator.
-
-    Class Attributes
-    ----------------
-    name : str
-        The name of the integrator.
-
-    supports_blackbox : bool
-        If True, then the integrator calls only ``system.matmul``,
-        ``system.matmul_data``, ``system.expect``, ``system.expect_data`` and
-        ``isconstant``, ``isoper`` or ``issuper``. This allows the solver using
-        the integrator to modify the system in creative ways. In particular,
-        the solver may modify the system depending on *both* the time ``t``
-        *and* the current ``state`` the system is being applied to.
-
-        If the integrator calls any other methods, set to False.
-
-    supports_time_dependent : bool
-        If True, then the integrator supports time dependent systems. If False,
-        ``supports_blackbox`` should usually be ``False`` too.
-
-    integrator_options : dict
-        A dictionary of options used by the integrator and their default
-        values. Once initiated, ``self.options`` will be a dict with the same
-        keys, not the full options object passed to the solver. Options' keys
-        included here will be supported by the :cls:SolverOdeOptions.
-    """
-    def set_state(self, t, state0, generator):
-        """
-        Set the state of the SODE solver.
-
-        Parameters
-        ----------
-        t : float
-            Initial time
-
-        state0 : qutip.Data
-            Initial state.
-
-        generator : numpy.random.generator
-            Random number generator.
-        """
-        raise NotImplementedError
-
-
-    def integrate(self, t, copy=True):
-        """
-        Evolve to t.
-
-        Before calling `integrate` for the first time, the initial state should
-        be set with `set_state`.
-
-        Parameters
-        ----------
-        t : float
-            Time to integrate to, should be larger than the previous time.
-
-        copy : bool [True]
-            Whether to return a copy of the state or the state itself.
-
-        Returns
-        -------
-        (t, state, noise) : (float, qutip.Data, np.ndarray)
-            The state of the solver at ``t``.
-        """
-        raise NotImplementedError
-
-
-    def mcstep(self, t, copy=True):
-        raise NotImplementedError
