+++ qutip/solver/sode/_sode.pyx
@@ -6,8 +6,6 @@ from collections import defaultdict
 cimport cython
 import numpy as np
 
-N_dws = defaultdict(lambda : 1)
-
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
@@ -20,6 +18,7 @@ cpdef Data euler(system, t, Data state, double dt, double[:, :] dW):
     By Peter E. Kloeden, Eckhard Platen
     """
     cdef int i
+
     a = system.drift(t, state)
     b = system.diffusion(t, state)
     new_state = _data.add(state, a, dt)
@@ -47,8 +46,10 @@ cpdef Data platen(system, t, Data state, double dt, double[:, :] dW):
     cdef double sqrt_dt_inv = 0.25 / sqrt_dt
     cdef double dw, dw2
 
-    d1 = _data.add(state, system.drift(t, state), dt)
-    d2 = system.diffusion(t, state)
+    cdef Data d1 = _data.add(state, system.drift(t, state), dt)
+    cdef list d2 = system.diffusion(t, state)
+    cdef Data Vt, out
+    cdef list Vp, Vm
 
     out = _data.mul(d1, 0.5)
     Vt = d1.copy()
@@ -76,3 +77,119 @@ cpdef Data platen(system, t, Data state, double dt, double[:, :] dW):
             out = _data.add(out, d2m[j], -dw2)
 
     return out
+
+
+@cython.boundscheck(False)
+@cython.wraparound(False)
+@cython.cdivision(True)
+cdef void explicit15(system, t, Data state, double dt, double[:, :] dW):
+    """
+    Chapter 11.2 Eq. (2.13)
+    Numerical Solution of Stochastic Differential Equations
+    By Peter E. Kloeden, Eckhard Platen
+    """
+    cdef int i, j, k
+    cdef double sqrt_dt = np.sqrt(dt)
+    cdef double sqrt_dt_inv = 1./sqrt_dt
+    cdef double ddz, ddw, ddd
+    cdef double[::1] dz, dw
+    dw = np.empty(system.num_collapse)
+    dz = np.empty(system.num_collapse)
+    for i in range(system.num_collapse):
+        dw[i] = dW[i, 0]
+        dz[i] = 0.5 *(dW[i, 0] + 1./np.sqrt(3) * noise[i, 1])
+
+    d1 = system.drift(t, state)
+    d2 = system.diffusion(t, state)
+    dd2 = system.diffusion(t+dt, state)
+    # Euler part
+    out = _data.add(state, d1)
+    for i in range(system.num_collapse):
+        out = _data.add(out, d2[:], dw[i])
+
+    V = _data.add(state, d1, 1./self.num_ops)
+
+    v2p = []
+    v2m = []
+    for i in range(system.num_collapse):
+        v2p.append(_data.add(V, d2[i], sqrt_dt))
+        v2m.append(_data.add(V, d2[i], -sqrt_dt))
+
+    p2p = []
+    p2m = []
+    for i in range(system.num_collapse):
+        d2p = system.diffusion(t, v2p[i])
+        d2m = system.diffusion(t, v2m[i])
+        ddw = (dw[i] * dw[i] - dt) * 0.25 * sqrt_dt_inv  # 1.0
+        out = _data.add(out, d2p[i], ddw)
+        out = _data.add(out, d2m[i], -ddw)
+        temp_p2p = []
+        temp_p2m = []
+        for j in range(system.num_collapse):
+            temp_p2p.append(v2p[i], d2p[j], sqrt_dt)
+            temp_p2m.append(v2p[i], d2p[j], -sqrt_dt)
+        p2p.append(temp_p2p)
+        p2m.append(temp_p2m)
+
+    out = _data.add(out, d1, -0.5*(self.num_ops))
+
+    for i in range(system.num_collapse):
+        ddz = dz[i] * 0.5 / sqrt_dt  # 1.5
+        ddd = 0.25 * (dw[i] * dw[i] / 3 - dt) * dw[i] / dt  # 1.5
+
+        d1p = system.drift(t + dt/self.num_ops, v2p[i])
+        d1m = system.drift(t + dt/self.num_ops, v2m[i])
+        d2p = system.diffusion(t, v2p[i])
+        d2m = system.diffusion(t, v2m[i])
+        d2pp = system.diffusion(t, p2p[i][i])
+        d2mm = system.diffusion(t, p2m[i][i])
+
+        out = _data.add(out, d1p, 0.25 + ddz)
+        out = _data.add(out, d1m, 0.25 - ddz)
+
+        out = _data.add(out, dd2[i], dw[i] - dz[i])
+        out = _data.add(out, d2[i], dz[i] - dw[i])
+
+        out = _data.add(out, d2pp[i], ddd)
+        out = _data.add(out, d2mm[i], ddd)
+        out = _data.add(out, d2p[i], ddd)
+        out = _data.add(out, d2m[i], ddd)
+
+        for j in range(system.num_collapse):
+            ddw = 0.5 * (dw[j] - dz[j])  # O(1.5)
+            out = _data.add(out, d2p[j], ddw)
+            out = _data.add(out, d2[j], -2*ddw)
+            out = _data.add(out, d2m[j], ddw)
+
+            if j > i:
+                ddw = 0.5 * (dw[i] * dw[j]) / sqrt_dt  # O(1.0)
+                out = _data.add(out, d2p[j], ddw)
+                out = _data.add(out, d2m[j], -ddw)
+
+                ddw = 0.25 * (dw[j] * dw[j] - dt) * dw[i] / dt  # O(1.5)
+                d2pp = self.diffusion(t, p2p[j][i])
+                d2mm = self.diffusion(t, p2m[j][i])
+
+                out = _data.add(out, d2pp[j], ddw)
+                out = _data.add(out, d2mm[j], -ddw)
+                out = _data.add(out, d2p[j], -ddw)
+                out = _data.add(out, d2m[j], ddw)s
+
+                for k in range(j+1, system.num_collapse):
+                    ddw = 0.5 * dw[i] * dw[j] * dw[k] / dt  # O(1.5)
+
+                    out = _data.add(out, d2pp[k], ddw)
+                    out = _data.add(out, d2mm[k], -ddw)
+                    out = _data.add(out, d2p[k], -ddw)
+                    out = _data.add(out, d2m[k], ddw)
+
+            if j < i:
+                ddw = 0.25 * (dw[j] * dw[j] - dt) * dw[i] / dt  # O(1.5)
+
+                d2pp = self.diffusion(t, p2p[j][i])
+                d2mm = self.diffusion(t, p2m[j][i])
+
+                out = _data.add(out, d2pp[j], ddw)
+                out = _data.add(out, d2mm[j], -ddw)
+                out = _data.add(out, d2p[j], -ddw)
+                out = _data.add(out, d2m[j], ddw)
